AI Chatbot/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pydocumentdb==2.3.3
Werkzeug==2.3.7
requests==2.31.0
